+++ lib/game_interface/interface_component.dart
@@ -1,5 +1,4 @@
 import 'package:bonfire/bonfire.dart';
-import 'package:bonfire/util/assets_loader.dart';
 import 'package:flutter/widgets.dart';
 
 /// Component used to add in your [GameInterface]
