+++ lib/joystick/joystick_directional.dart
@@ -1,7 +1,6 @@
 import 'dart:math';
 
 import 'package:bonfire/bonfire.dart';
-import 'package:bonfire/util/assets_loader.dart';
 import 'package:flutter/material.dart';
 
 class JoystickDirectional {
