+++ lib/npc/ally/rotation_ally.dart
@@ -1,5 +1,4 @@
 import 'package:bonfire/bonfire.dart';
-import 'package:bonfire/util/assets_loader.dart';
 
 ///
 /// Created by
