+++ lib/npc/rotation_npc.dart
@@ -1,5 +1,4 @@
 import 'package:bonfire/bonfire.dart';
-import 'package:bonfire/util/assets_loader.dart';
 
 ///
 /// Created by
