+++ lib/objects/animated_object_once.dart
@@ -2,7 +2,6 @@ import 'dart:async';
 import 'dart:ui';
 
 import 'package:bonfire/bonfire.dart';
-import 'package:bonfire/util/assets_loader.dart';
 
 class AnimatedObjectOnce extends GameComponent with UseAssetsLoader, Lighting {
   final VoidCallback? onFinish;
