+++ lib/util/controlled_update_animation.dart
@@ -1,7 +1,6 @@
 import 'dart:ui';
 
 import 'package:bonfire/bonfire.dart';
-import 'package:bonfire/util/assets_loader.dart';
 
 class ControlledUpdateAnimation {
   bool _alreadyUpdate = false;
