+++ lib/util/direction_animations/simple_direction_animation.dart
@@ -3,7 +3,6 @@ import 'dart:ui';
 
 import 'package:bonfire/base/bonfire_game_interface.dart';
 import 'package:bonfire/bonfire.dart';
-import 'package:bonfire/util/assets_loader.dart';
 
 /// Class responsible to manager animation on `SimplePlayer` and `SimpleEnemy`
 class SimpleDirectionAnimation {
